+++ pyannote/audio/tasks/embedding/mixins.py
@@ -75,7 +75,7 @@ class SupervisedRepresentationLearningTaskMixin:
     def batch_size(self, batch_size: int):
         self.batch_size_ = batch_size
 
-    def setup(self):
+    def setup(self, stage=None):
         # loop over the training set, remove annotated regions shorter than
         # chunk duration, and keep track of the reference annotations, per class.
 
@@ -118,12 +118,6 @@ class SupervisedRepresentationLearningTaskMixin:
             min_duration=self.min_duration,
             classes=sorted(self._train),
         )
-
-        if not self.has_validation:
-            return
-
-        if isinstance(self.protocol, SpeakerVerificationProtocol):
-            self._validation = list(self.protocol.development_trial())
 
     def default_metric(
         self,
@@ -250,9 +244,13 @@ class SupervisedRepresentationLearningTaskMixin:
 
         return {"loss": loss}
 
+    def prepare_validation(self, prepared_dict: Dict):
+        if isinstance(self.protocol, SpeakerVerificationProtocol):
+            prepared_dict["validation"] = list(self.protocol.development_trial())
+
     def val__getitem__(self, idx):
         if isinstance(self.protocol, SpeakerVerificationProtocol):
-            trial = self._validation[idx]
+            trial = self.prepared_data["validation"][idx]
 
             data = dict()
             for idx in [1, 2]:
@@ -281,7 +279,7 @@ class SupervisedRepresentationLearningTaskMixin:
 
     def val__len__(self):
         if isinstance(self.protocol, SpeakerVerificationProtocol):
-            return len(self._validation)
+            return len(self.prepared_data["validation"])
 
         elif isinstance(self.protocol, SpeakerDiarizationProtocol):
             return 0