+++ tests/tasks/test_reproducibility.py
@@ -3,7 +3,7 @@ from lightning.pytorch import seed_everything
 from pyannote.database import FileFinder, get_protocol
 
 from pyannote.audio.models.segmentation.debug import SimpleSegmentationModel
-from pyannote.audio.tasks import MultiLabelSegmentation, VoiceActivityDetection
+from pyannote.audio.tasks import VoiceActivityDetection
 
 
 def setup_tasks(task):
@@ -16,7 +16,8 @@ def setup_tasks(task):
 
 def create_dl(model, task):
     m = model(task=task)
-    m.setup("fit")
+    m.prepare_data()
+    m.setup()
     return task.train_dataloader()
 
 
@@ -31,35 +32,32 @@ def get_next5(dl):
 def test_seeding_ensures_data_loaders():
     "Setting a global seed for the dataloaders ensures that we get data back in the same order"
 
-    for task in [VoiceActivityDetection, MultiLabelSegmentation]:
+    seed_everything(1)
+    protocol, vad = setup_tasks(VoiceActivityDetection)
+    dl = create_dl(SimpleSegmentationModel, vad)
+    last5a = get_next5(dl)
 
-        seed_everything(1)
-        protocol, vad = setup_tasks(task)
-        dl = create_dl(SimpleSegmentationModel, vad)
-        last5a = get_next5(dl)
+    seed_everything(1)
+    protocol, vad = setup_tasks(VoiceActivityDetection)
+    dl = create_dl(SimpleSegmentationModel, vad)
+    last5b = get_next5(dl)
 
-        seed_everything(1)
-        protocol, vad = setup_tasks(task)
-        dl = create_dl(SimpleSegmentationModel, vad)
-        last5b = get_next5(dl)
-
-        for i in range(len(last5b)):
-            assert torch.equal(last5a[i]["X"], last5b[i]["X"])
+    for i in range(len(last5b)):
+        assert torch.equal(last5a[i]["X"], last5b[i]["X"])
 
 
 def test_different_seeds():
     "Changing the global seed will change the order of the data that loads"
 
-    for task in [VoiceActivityDetection, MultiLabelSegmentation]:
-        protocol, vad = setup_tasks(task)
-        seed_everything(4)
-        dl = create_dl(SimpleSegmentationModel, vad)
-        last5a = get_next5(dl)
+    protocol, vad = setup_tasks(VoiceActivityDetection)
+    seed_everything(4)
+    dl = create_dl(SimpleSegmentationModel, vad)
+    last5a = get_next5(dl)
 
-        protocol, vad = setup_tasks(task)
-        seed_everything(5)
-        dl = create_dl(SimpleSegmentationModel, vad)
-        last5b = get_next5(dl)
+    protocol, vad = setup_tasks(VoiceActivityDetection)
+    seed_everything(5)
+    dl = create_dl(SimpleSegmentationModel, vad)
+    last5b = get_next5(dl)
 
-        for i in range(5):
-            assert not torch.equal(last5a[i]["X"], last5b[i]["X"])
+    for i in range(5):
+        assert not torch.equal(last5a[i]["X"], last5b[i]["X"])